+++ Modules/Microsoft365DSC/Modules/M365DSCAgent.psm1
@@ -5,7 +5,6 @@ function Test-M365DSCAgent
     param(
 
     )
-    $InformationPreference = 'Continue'
     #region Telemetry
     $data = [System.Collections.Generic.Dictionary[[String], [String]]]::new()
     $data.Add("Event", "TestAgent")