+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+* EXOAvailabilityConfig
+  * Fixed an issue with the Test-TargetResource where if a full
+    username (with '@') was specified for the OrgWideAccount, it
+    would always return false;
 * EXOMobileDeviceMailboxPolicy
   * Fixed an error where if no MinPasswordLength was specified
     the Set-TargetResource threw n error trying to create a new